bugs fixed
