+++ src/Components/HeaderGridImages/HeaderGridImages.module.css
@@ -2,13 +2,10 @@
     background-color: var(--main-dark-color);
     border-radius: 12px;
     padding: 1rem;
-    position: absolute;
     display: grid;
     grid-template-columns: repeat(2, 1fr);
     grid-template-rows: repeat(3, 1fr);
-    gap: 30px 10px;
-    right: 0px;
-    top: 50px;
+    gap: 10px;
 }
 .header_images img {
     max-width: 200px;
@@ -47,6 +44,11 @@
 .forth_image {
     grid-column: 2 / 3;
     grid-row: 3 / 4;
+}
+.forth_image video {
+    width: 200px;
+    object-fit: cover;
+    border-radius: 12px;
 }
 
 .fivth_image {
