+++ src/Components/HomeHeader/HomeHeader.module.css
@@ -45,25 +45,17 @@
     color: var(--main-color);
 }
 .bg_video {
-    max-width: 200px;
-    position: relative;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 0px;
-    top: 15px;
+    width: 200px;
+    height: 500px;
+    padding-top: 20px;
+    position: absolute;
+    right: 15px;
+    top: 40%;
 }
 
 .bg_video video {
-    width: 100%;
-    height: 100%;
-}
-
-.bg_video video {
-    width: 100%;
-    height: 100%;
+    width: 200px;
     border-radius: 20px;
-    margin: 100px 0;
     z-index: 9999999999;
     box-shadow: rgba(0, 0, 0, 0.25) 0px 54px 55px,
         rgba(0, 0, 0, 0.12) 0px -12px 30px, rgba(0, 0, 0, 0.12) 0px 4px 6px,
